+++ mediafile/__init__.py
@@ -60,6 +60,7 @@ from .storage import (
     MP3DescStorageStyle,
     MP3ImageStorageStyle,
     MP3ListDescStorageStyle,
+    MP3ListPeopleStorageStyle,
     MP3ListStorageStyle,
     MP3PeopleStorageStyle,
     MP3SlashPackStorageStyle,
@@ -402,30 +403,33 @@ class MediaFile:
     )
     genre = genres.single_field()
 
-    lyricist = MediaField(
-        MP3StorageStyle("TEXT"),
-        MP4StorageStyle("----:com.apple.iTunes:LYRICIST"),
-        StorageStyle("LYRICIST"),
+    lyricists = ListMediaField(
+        MP3ListStorageStyle("TEXT"),
+        MP4ListStorageStyle("----:com.apple.iTunes:LYRICIST"),
+        ListStorageStyle("LYRICIST"),
         ASFStorageStyle("WM/Writer"),
     )
-    composer = MediaField(
-        MP3StorageStyle("TCOM"),
-        MP4StorageStyle("\xa9wrt"),
-        StorageStyle("COMPOSER"),
+    lyricist = lyricists.single_field()
+    composers = ListMediaField(
+        MP3ListStorageStyle("TCOM"),
+        MP4ListStorageStyle("\xa9wrt"),
+        ListStorageStyle("COMPOSER"),
         ASFStorageStyle("WM/Composer"),
     )
+    composer = composers.single_field()
     composer_sort = MediaField(
         MP3StorageStyle("TSOC"),
         MP4StorageStyle("soco"),
         StorageStyle("COMPOSERSORT"),
         ASFStorageStyle("WM/Composersortorder"),
     )
-    arranger = MediaField(
-        MP3PeopleStorageStyle("TIPL", involvement="arranger"),
-        MP4StorageStyle("----:com.apple.iTunes:Arranger"),
-        StorageStyle("ARRANGER"),
+    arrangers = ListMediaField(
+        MP3ListPeopleStorageStyle("TIPL", involvement="arranger"),
+        MP4ListStorageStyle("----:com.apple.iTunes:Arranger"),
+        ListStorageStyle("ARRANGER"),
         ASFStorageStyle("beets/Arranger"),
     )
+    arranger = arrangers.single_field()
 
     grouping = MediaField(
         MP3StorageStyle("TIT1"),